Massive cleanup commit.   - Fixed some coding conventions that were bothering me.   - Changed instances of single character strings to single characters.   - Changed some larger if/elseif/else assignment blocks to use static arrays to assign values.

+++ src/dataop_coproc.c
@@ -24,8 +24,9 @@
 
 #define ADIS_OPCODE(_op)        ((_op & 0x00F00000) >> 20)
 
-void dataop_coproc_instr(uint32_t op) {
-    uint8_t coproc_info;
+void dataop_coproc_instr(uint32_t op)
+{
+    uint32_t coproc_info;
     char *cond;
 
     cond = get_condition_string(op);
